# xswagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LAARTESA API",
    "version": "1.0.0",
    "description": "Documentación de la API para WEB APP LA ARTESA"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor Local"
    },
    {
      "url": "https://f686-160-173-184-21.ngrok-free.app",
      "description": "Servidor Ngrok"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ingrese el token JWT en el formato: Bearer <token>"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "mail": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico del usuario",
            "example": "usuario@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Contraseña del usuario",
            "example": "contraseña123"
          }
        },
        "required": ["mail", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Login exitoso"
          },
          "data": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              },
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "mail": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "rol_id": {
                    "type": "integer",
                    "example": 2
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registrar un nuevo usuario",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "mail", "password"],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "Nombre del usuario"
                  },
                  "mail": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 100,
                    "description": "Correo electrónico del usuario (debe ser único)"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 6,
                    "maxLength": 255,
                    "description": "Contraseña del usuario (mínimo 6 caracteres)"
                  },
                  "rol_id": {
                    "type": "integer",
                    "default": 2,
                    "description": "ID del rol del usuario (2 = usuario normal por defecto)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Iniciar sesión y obtener token JWT",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
  }
}